format code

+++ src/main/java/com/nthuy/demo_erm/service/Impl/AttributeGroupServiceImpl.java
@@ -1,33 +1,30 @@
-package com.nthuy.demo_erm.service;
+package com.nthuy.demo_erm.service.Impl;
 
 import com.nthuy.demo_erm.config.AttributeGroupSpecification;
-import com.nthuy.demo_erm.config.ClassifyReasonSpecification;
 import com.nthuy.demo_erm.constant.EnumTypeAttributeGroup;
 import com.nthuy.demo_erm.dto.*;
 import com.nthuy.demo_erm.entity.AttributeGroupEntity;
-import com.nthuy.demo_erm.entity.ClassifyReasonEntity;
-import com.nthuy.demo_erm.entity.RiskCategoryEntity;
-import com.nthuy.demo_erm.entity.SystemEntity;
 import com.nthuy.demo_erm.exception.BadRequestValidationException;
+import com.nthuy.demo_erm.exception.IdInvalidException;
+import com.nthuy.demo_erm.exception.NameExisted;
 import com.nthuy.demo_erm.exception.TypeAttributeGroupValidException;
 import com.nthuy.demo_erm.mapper.AttributeGroupMapper;
 import com.nthuy.demo_erm.repository.AttributeGroupRepository;
+import com.nthuy.demo_erm.service.AttributeGroupService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
-import java.util.HashSet;
+
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
+
 
 @Service
-public class AttributeGroupServiceImpl implements AttributeGroupService{
+public class AttributeGroupServiceImpl implements AttributeGroupService {
 
-   private final AttributeGroupRepository attributeGroupRepository;
-   private final AttributeGroupMapper attributeGroupMapper;
+    private final AttributeGroupRepository attributeGroupRepository;
+    private final AttributeGroupMapper attributeGroupMapper;
 
     public AttributeGroupServiceImpl(AttributeGroupRepository attributeGroupRepository, AttributeGroupMapper attributeGroupMapper) {
         this.attributeGroupRepository = attributeGroupRepository;
@@ -35,24 +32,26 @@ public class AttributeGroupServiceImpl implements AttributeGroupService{
     }
 
 
-    @Override
-    public boolean nameExists(String userName) {
+    private boolean nameExists(String userName) {
         return this.attributeGroupRepository.existsByName(userName);
     }
 
-    @Override
-    public boolean existsById(Long id) {
+
+    private boolean existsById(Long id) {
         return this.attributeGroupRepository.existsById(id);
     }
 
     @Override
-    public Long handleCreateAttributeGroup(AttributeGroupDTO dto) {
+    public Long handleCreateAttributeGroup(AttributeGroupDTO dto) throws NameExisted {
+        if (this.nameExists(dto.getName())) {
+            throw new NameExisted("tên đã có");
+        }
         AttributeGroupEntity entity = attributeGroupMapper.toEntity(dto);
-        if (entity.getType()!=null){
+        if (entity.getType() != null) {
             System.out.println("Entity trước khi lưu: " + entity);
             AttributeGroupEntity savedEntity = attributeGroupRepository.save(entity);
             return savedEntity.getId();
-        }else {
+        } else {
             entity.setType(EnumTypeAttributeGroup.BUSINESS);
             System.out.println("Entity trước khi lưu: " + entity);
             AttributeGroupEntity savedEntity = attributeGroupRepository.save(entity);
@@ -70,20 +69,27 @@ public class AttributeGroupServiceImpl implements AttributeGroupService{
 
     @Override
     public void handleDeleteAttributeGroup(Long id) {
+        if (this.existsById(id)) {
+            throw new IdInvalidException("Id không có" + id);
+        }
+
         AttributeGroupDTO dto = this.handleGetAttributeGroupById(id);
-        if (!dto.getType().equals(EnumTypeAttributeGroup.SYSTEM)){
+        if (!dto.getType().equals(EnumTypeAttributeGroup.SYSTEM)) {
             this.attributeGroupRepository.deleteById(dto.getId());
-        }else throw new  TypeAttributeGroupValidException("SYSTEM không thể xóa");
+        } else throw new TypeAttributeGroupValidException("SYSTEM không thể xóa");
 
     }
 
     @Override
-    public Long handleUpdateAttributeGroup(AttributeGroupDTO dto) {
+    public Long handleUpdateAttributeGroup(AttributeGroupDTO dto) throws NameExisted {
+        if (this.nameExists(dto.getName())) {
+            throw new NameExisted("tên đã có");
+        }
         AttributeGroupEntity entity = attributeGroupRepository.findById(dto.getId())
                 .orElseThrow(() -> new BadRequestValidationException(
                         "Nhóm thuộc tính với ID " + dto.getId() + " không tồn tại"));
-        attributeGroupMapper.updateEntityFromDto(dto,entity);
-        if (entity.getType().equals(EnumTypeAttributeGroup.SYSTEM)){
+        attributeGroupMapper.updateEntityFromDto(dto, entity);
+        if (entity.getType().equals(EnumTypeAttributeGroup.SYSTEM)) {
             throw new TypeAttributeGroupValidException("System không thể chỉnh sửa");
         }
         return this.attributeGroupRepository.save(entity).getId();
