+++ src/main/java/com/nthuy/demo_erm/service/Impl/AttributeServiceImpl.java
@@ -1,18 +1,19 @@
-package com.nthuy.demo_erm.service;
+package com.nthuy.demo_erm.service.Impl;
 
 import com.nthuy.demo_erm.constant.EnumAttributeDisplayType;
 import com.nthuy.demo_erm.dto.AttributeDTO;
-import com.nthuy.demo_erm.dto.AttributeGroupDTO;
 import com.nthuy.demo_erm.dto.AttributeValueDTO;
 import com.nthuy.demo_erm.dto.ResultPaginationDTO;
 import com.nthuy.demo_erm.entity.AttributeEntity;
 import com.nthuy.demo_erm.entity.AttributeValueEntity;
-import com.nthuy.demo_erm.entity.ClassifyReasonEntity;
 import com.nthuy.demo_erm.exception.BadRequestValidationException;
+import com.nthuy.demo_erm.exception.IdInvalidException;
+import com.nthuy.demo_erm.exception.NameExisted;
 import com.nthuy.demo_erm.mapper.AttributeMapper;
 import com.nthuy.demo_erm.mapper.AttributeValueMapper;
 import com.nthuy.demo_erm.repository.AttributeRepository;
 import com.nthuy.demo_erm.repository.AttributeValueRepository;
+import com.nthuy.demo_erm.service.AttributeService;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,7 +25,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
-public class AttributeServiceImpl implements AttributeService{
+public class AttributeServiceImpl implements AttributeService {
 
     private final AttributeRepository attributeRepository;
     private final AttributeValueRepository attributeValueRepository;
@@ -39,18 +40,18 @@ public class AttributeServiceImpl implements AttributeService{
     }
 
 
-    @Override
-    public boolean nameExists(String name) {
+
+    private boolean nameExists(String name) {
         return this.attributeRepository.existsByName(name);
     }
 
-    @Override
-    public boolean existsById(Long id) {
+
+    private boolean existsById(Long id) {
         return this.attributeRepository.existsById(id);
     }
-
+    @Override
     @Transactional
-    public Long handleCreateAttribute(AttributeDTO dto) {
+    public Long handleCreateAttribute(AttributeDTO dto) throws NameExisted {
         // gọi validate riêng
         validateAttribute(dto);
 
@@ -96,21 +97,32 @@ public class AttributeServiceImpl implements AttributeService{
 
     @Override
     public void handleDeleteAttribute(Long id) {
+        if (this.existsById(id)) {
+            throw new IdInvalidException("ID " + id + " không có");
+        }
         this.attributeRepository.deleteById(id);
     }
 
     @Override
     public Long handleUpdateAttribute(AttributeDTO dto) {
-        return 0L;
+        AttributeEntity entity = this.attributeRepository.findById(dto.getId())
+                .orElseThrow(() -> new BadRequestValidationException(
+                "Thuộc tính với ID " + dto.getId() + " không tồn tại"));
+        attributeMapper.updateEntityFromDto(dto,entity);
+        return this.attributeRepository.save(entity).getId();
     }
 
     @Override
     public ResultPaginationDTO<AttributeDTO> handleGetAttribute(String code, String name, Boolean isActive, Pageable pageable) {
         return null;
     }
-    public void validateAttribute(AttributeDTO dto) {
+
+    private void validateAttribute(AttributeDTO dto) throws NameExisted {
         if (dto == null) {
             throw new IllegalArgumentException("Payload không được null");
+        }
+        if (this.nameExists(dto.getName()))   {
+            throw new NameExisted("Tên đã có");
         }
 
         EnumAttributeDisplayType displayType = dto.getDisplayType() == null
