+++ src/main/java/com/nthuy/demo_erm/service/Impl/ReasonServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nthuy.demo_erm.service;
+package com.nthuy.demo_erm.service.Impl;
 
 
 import com.nthuy.demo_erm.config.ReasonSpecification;
@@ -9,9 +9,13 @@ import com.nthuy.demo_erm.dto.ResultPaginationDTO;
 import com.nthuy.demo_erm.entity.ReasonEntity;
 import com.nthuy.demo_erm.entity.SystemEntity;
 import com.nthuy.demo_erm.exception.BadRequestValidationException;
+import com.nthuy.demo_erm.exception.IdInvalidException;
+import com.nthuy.demo_erm.exception.NameExisted;
 import com.nthuy.demo_erm.mapper.ReasonMapper;
 import com.nthuy.demo_erm.repository.ReasonRepository;
 import com.nthuy.demo_erm.repository.SystemRepository;
+import com.nthuy.demo_erm.service.ReasonService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -24,101 +28,85 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class ReasonServiceImpl implements ReasonService {
 
     private final ReasonRepository reasonRepository;
     private final ReasonMapper reasonMapper;
     private final SystemRepository systemRepository;
 
-    public ReasonServiceImpl(ReasonRepository reasonRepository, ReasonMapper reasonMapper, SystemRepository systemRepository) {
-        this.reasonRepository = reasonRepository;
-        this.reasonMapper = reasonMapper;
-        this.systemRepository = systemRepository;
-    }
 
-
-    public boolean nameExists(String userName) {
+    private boolean nameExists(String userName) {
         return this.reasonRepository.existsByName(userName);
     }
 
-
-    public boolean existsById(Long id) {
+    private boolean existsById(Long id) {
         return this.reasonRepository.existsById(id);
     }
 
-    public Long handleCreateClassifyReason(ReasonDTO dto) {
+    @Override
+    public Long handleCreateClassifyReason(ReasonDTO dto) throws NameExisted {
+        if (this.nameExists(dto.getName())) {
+            throw new NameExisted("Name đã có");
+        }
         ReasonEntity entity = reasonMapper.toEntity(dto);
-
         Set<SystemEntity> systemEntities;
         if (dto.getSystems() != null && !dto.getSystems().isEmpty()) {
             // Nếu DTO có truyền systems thì lấy theo danh sách đó
-            systemEntities = dto.getSystems().stream()
-                    .map(systemDTO -> systemRepository.findById(systemDTO.getId())
-                            .orElseThrow(() -> new RuntimeException("System not found with id: " + systemDTO.getId())))
-                    .collect(Collectors.toSet());
+            systemEntities = dto.getSystems().stream().map(systemDTO -> systemRepository.findById(systemDTO.getId()).orElseThrow(() -> new RuntimeException("System not found with id: " + systemDTO.getId()))).collect(Collectors.toSet());
         } else {
             // Nếu không truyền thì mặc định lấy system có id = 1 và id = 2
             systemEntities = new HashSet<>(systemRepository.findAllById(Arrays.asList(1L, 2L)));
         }
-
         entity.setSystemEntitiesReason(systemEntities);
-
-// In ra để kiểm tra
         System.out.println("Entity trước khi lưu: " + entity);
-
-// Lưu xuống DB
         ReasonEntity savedEntity = reasonRepository.save(entity);
 
-// Trả về ID
         return savedEntity.getId();
     }
 
+    @Override
     public ReasonDTO handleGetReasonById(Long id) {
 
-        ReasonEntity reason = this.reasonRepository.findById(id)
-                .orElseThrow(() -> new BadRequestValidationException("Thẻ bảo hiểm với ID " + id + " không tồn tại"));
+        ReasonEntity reason = this.reasonRepository.findById(id).orElseThrow(() -> new BadRequestValidationException("Thẻ bảo hiểm với ID " + id + " không tồn tại"));
         return reasonMapper.toDto(reason);
 
     }
 
-
+    @Override
     public void handleDeleteReason(Long id) {
+        if (this.existsById(id)) {
+            throw new IdInvalidException("Id không có" + id);
+        }
         this.reasonRepository.deleteById(id);
     }
 
-
-    public Long handleUpdateReason(ReasonDTO dto) {
+    @Override
+    public Long handleUpdateReason(ReasonDTO dto) throws NameExisted {
         // Lấy entity cũ từ DB
-        ReasonEntity reason = reasonRepository.findById(dto.getId())
-                .orElseThrow(() -> new BadRequestValidationException(
-                        "Phân loại nguyên nhân với ID " + dto.getId() + " không tồn tại"));
-        reasonMapper.updateEntityFromDto(dto,reason);
+        if (this.nameExists(dto.getName())) {
+            throw new NameExisted("Name đã có");
+        }
+        ReasonEntity reason = reasonRepository.findById(dto.getId()).orElseThrow(() -> new BadRequestValidationException("Phân loại nguyên nhân với ID " + dto.getId() + " không tồn tại"));
+        reasonMapper.updateEntityFromDto(dto, reason);
         Set<SystemEntity> systemEntities;
 
         if (dto.getSystems() != null && !dto.getSystems().isEmpty()) {
             // Nếu DTO có truyền systems thì lấy theo danh sách đó
-            systemEntities = dto.getSystems().stream()
-                    .map(systemDTO -> systemRepository.findById(systemDTO.getId())
-                            .orElseThrow(() -> new RuntimeException("System not found with id: " + systemDTO.getId())))
-                    .collect(Collectors.toSet());
+            systemEntities = dto.getSystems().stream().map(systemDTO -> systemRepository.findById(systemDTO.getId()).orElseThrow(() -> new RuntimeException("System not found with id: " + systemDTO.getId()))).collect(Collectors.toSet());
         } else {
             // Nếu không truyền thì mặc định lấy system có id = 1 và id = 2
             systemEntities = new HashSet<>(systemRepository.findAllById(Arrays.asList(1L, 2L)));
         }
-
         reason.setSystemEntitiesReason(systemEntities);
+
         return reasonRepository.save(reason).getId();
     }
 
-    public ResultPaginationDTO<ReasonDTO> handleGetReason(
-            String code, String name, List<Long> systemIds, Boolean isActive, EnumTypeReason type, Pageable pageable) {
+    @Override
+    public ResultPaginationDTO<ReasonDTO> handleGetReason(String code, String name, List<Long> systemIds, Boolean isActive, EnumTypeReason type, Pageable pageable) {
 
-        Specification<ReasonEntity> spec = Specification
-                .where(ReasonSpecification.hasCode(code))
-                .and(ReasonSpecification.hasName(name))
-                .and(ReasonSpecification.hasSystemIdIn(systemIds))
-                .and(ReasonSpecification.hasType(type))
-                .and(ReasonSpecification.hasIsActive(isActive));
+        Specification<ReasonEntity> spec = Specification.where(ReasonSpecification.hasCode(code)).and(ReasonSpecification.hasName(name)).and(ReasonSpecification.hasSystemIdIn(systemIds)).and(ReasonSpecification.hasType(type)).and(ReasonSpecification.hasIsActive(isActive));
 
         Page<ReasonEntity> pageResult = reasonRepository.findAll(spec, pageable);
